Implement alignment annotation mapping in API
- Introduced a new private function `_alignment_annotation_mapping` in `annotations.py` to handle the mapping of alignment annotations and target annotations, generating unique IDs for segments.
- Updated `_create_aligned_text` function in `instances.py` to utilize the new mapping function, streamlining the process of creating aligned manifestations and improving code clarity by removing redundant ID generation logic.
